+++ gradio/components.py
@@ -46,6 +46,7 @@ from gradio.events import (
     Submittable,
     Uploadable,
 )
+from gradio.interpretation import Interpretable
 from gradio.layouts import Column, Form, Row
 from gradio.processing_utils import TempFileManager
 from gradio.serializing import (
@@ -186,8 +187,6 @@ class IOComponent(Component, Serializable):
         if callable(load_fn):
             self.load_event = self.attach_load_event(load_fn, every)
 
-        self.set_interpret_parameters()
-
     def get_config(self):
         return {
             "label": self.label,
@@ -195,38 +194,6 @@ class IOComponent(Component, Serializable):
             "interactive": self.interactive,
             **super().get_config(),
         }
-
-    def set_interpret_parameters(self):
-        """
-        Set any parameters for interpretation.
-        """
-        return self
-
-    def get_interpretation_neighbors(self, x: Any) -> Tuple[List, Dict, bool]:
-        """
-        Generates values similar to input to be used to interpret the significance of the input in the final output.
-        Parameters:
-            x: Input to interface
-        Returns: (neighbor_values, interpret_kwargs, interpret_by_removal)
-            neighbor_values: Neighboring values to input x to compute for interpretation
-            interpret_kwargs: Keyword arguments to be passed to get_interpretation_scores
-            interpret_by_removal: If True, returned neighbors are values where the interpreted subsection was removed. If False, returned neighbors are values where the interpreted subsection was modified to a different value.
-        """
-        return [], {}, True
-
-    def get_interpretation_scores(
-        self, x: Any, neighbors: List[Any], scores: List[float], **kwargs
-    ) -> List:
-        """
-        Arrange the output values from the neighbors into interpretation scores for the interface to render.
-        Parameters:
-            x: Input to interface
-            neighbors: Neighboring values to input x used for interpretation.
-            scores: Output value corresponding to each neighbor in neighbors
-        Returns:
-            Arrangement of interpretation scores for interfaces to render.
-        """
-        return []
 
     def generate_sample(self) -> Any:
         """
@@ -275,7 +242,7 @@ class FormComponent:
 
 @document("change", "submit", "blur", "style")
 class Textbox(
-    FormComponent, Changeable, Submittable, Blurrable, IOComponent, SimpleSerializable
+    FormComponent, Changeable, Submittable, Blurrable, IOComponent, SimpleSerializable, Interpretable
 ):
     """
     Creates a textarea for user to enter string input or display string output.
@@ -324,7 +291,6 @@ class Textbox(
         self.lines = lines
         self.max_lines = max_lines if type == "text" else 1
         self.placeholder = placeholder
-        self.interpret_by_tokens = True
         IOComponent.__init__(
             self,
             label=label,
@@ -336,6 +302,7 @@ class Textbox(
             value=value,
             **kwargs,
         )
+        Interpretable.__init__(self, interpret_by_tokens=True)
         self.cleared_value = ""
         self.test_input = value
         self.type = type
